a lot of changes: yaml, dataset save, KENN, clause generation
changes in yaml management, now yaml are modular and there exist default yamls; KENNDataset removed, its parameters are now in KENNClassifier; the type2id can be created based on the dataset or given as a list of types; clauses in KENN are generated directly from the KENNClassifier and no more from the KENNDataset; type2id is expected to be linked in the BaseNetwork as parameter. A list of types is saved when the type2id is generated. a default folder for the typelist and tokenized dataset is supported